Change data type to match assigned one.

+++ crnlib/crn_vector.cpp
@@ -21,7 +21,7 @@ namespace crnlib
       if (m_capacity >= min_new_capacity)
          return true;
 
-      size_t new_capacity = min_new_capacity;
+      uint new_capacity = min_new_capacity;
       if ((grow_hint) && (!math::is_power_of_2(new_capacity)))
          new_capacity = math::next_pow2(new_capacity);
